+++ doc/CHANGELOG.rst
@@ -3,8 +3,9 @@ Changelog
 
 0.5.1x (MM/DD/YYYY)
 -------------------
-- Extended to support pysam v0.14
+- Extended to support pysam v0.13, v0.14
 - Adjusted so warning instead of error when no families found
+- Added optional command line arg to sort filters results by name instead of count 
 - Substantial refactors to clarify implementation
 
 0.5.1 (9/8/2017)